+++ components/Docs/PageActions.tsx
@@ -13,13 +13,13 @@ export const PageActions: React.FunctionComponent<IPageActionsProps> = ({ page }
   }
 
   return (
-    <div className={`absolute right-0`}>
+    <div className={`mt-16 mb-8`}>
       <a
-        className={`flex items-center text-whisper-500 hover:text-whisper-900 -mt-4`}
+        className={`flex items-center text-whisper-900 hover:text-whisper-500`}
         href={`${Extension.githubDocs}/edit/${process.env.NEXT_PUBLIC_VERCEL_ENV === "production" ? "main" : "dev"}/content/docs/${page.fileName}.md`}
         target="_blank"
         rel={`noopener noreferrer`}>
-        <PencilIcon className={`w-4 h-4 mr-2`} /> <span>Edit</span>
+        <PencilIcon className={`w-4 h-4 mr-2`} /> <span>Edit page</span>
       </a>
     </div>
   );