+++ lib/publicUrl.ts
@@ -1,9 +1,9 @@
-
-
 export function publicUrl() {
   if (process.env.NEXT_PUBLIC_VERCEL_ENV === "production") {
     return `https://frontmatter.codes`;
+  } else if (process.env.NEXT_PUBLIC_VERCEL_ENV === "preview") {
+    return `https://beta.frontmatter.codes`;
   } else {
-    return `https://${process.env.VERCEL_URL}`;
+    return `http://localhost:3000`;
   }
 }